Add inverse to benchmark
